+++ Cargo.toml
@@ -5,7 +5,7 @@ cargo-features = ["profile-rustflags"]
 
 [package]
 name = "pexrc"
-version = "0.15.0"
+version = "0.16.0"
 edition = { workspace = true }
 publish = false
 
@@ -210,6 +210,7 @@ url = "2.5"
 # does not have a feature to activate the dep.
 version-ranges = "*"
 walkdir = "2.5"
+windows-registry = "0.6"
 which = "8.0"
 xz2 = "0.1"
 zip = { version = "8.6", default-features = false, features = ["chrono", "deflate", "zstd"] }
@@ -253,6 +254,7 @@ python-proxy = { path = "crates/python-proxy" }
 rayon = { workspace = true }
 request = { path = "crates/request" }
 scripts = { path = "crates/scripts", features = ["embedded"] }
+serde_json = { workspace = true }
 sha2 = { workspace = true }
 strum = { workspace = true }
 target = { path = "crates/target" }
