+++ crates/interpreter/Cargo.toml
@@ -28,6 +28,9 @@ time = { workspace = true }
 url = { workspace = true }
 which = { workspace = true }
 
+[target.'cfg(windows)'.dependencies]
+windows-registry = { workspace = true }
+
 [dev-dependencies]
 anyhow = { workspace = true }
 pep440_rs = { workspace = true }
