+++ src/commands/mod.rs
@@ -6,10 +6,12 @@ mod extract;
 pub mod info;
 mod inject;
 mod platform;
+mod python;
 mod script;
 
 pub use build::Build;
 pub use extract::Extract;
 pub use inject::Inject;
 pub use platform::Platform;
+pub use python::Python;
 pub use script::Script;
